--- client/web-sveltekit/src/routes/[...repo=reporev]/-/commit/[...revspec]/page.gql
@@ -1,41 +1,0 @@
-query CommitPage_CommitQuery($repoName: String!, $revspec: String!) {
-    repository(name: $repoName) {
-        id
-        commit(rev: $revspec) {
-            id
-            oid
-            parents {
-                id
-                oid
-                abbreviatedOID
-                canonicalURL
-            }
-            externalURLs {
-                url
-                serviceKind
-            }
-            ...Commit
-        }
-    }
-}
-
-query CommitPage_DiffQuery($repoName: String!, $base: String, $head: String, $first: Int, $after: String) {
-    repository(name: $repoName) {
-        id
-        comparison(base: $base, head: $head) {
-            fileDiffs(first: $first, after: $after) {
-                ...CommitPage_DiffConnection
-            }
-        }
-    }
-}
-
-fragment CommitPage_DiffConnection on FileDiffConnection {
-    nodes {
-        ...FileDiff_Diff
-    }
-    pageInfo {
-        endCursor
-        hasNextPage
-    }
-}
